Added timeout to MalcolmDevice calls. This reduces risk of hangs.
