+++ org.eclipse.scanning.example/src/org/eclipse/scanning/example/malcolm/EPICSv4ExampleModel.java
@@ -13,7 +13,7 @@ package org.eclipse.scanning.example.malcolm;
 
 import org.eclipse.scanning.api.device.models.MalcolmModel;
 
-public class ExampleMalcolmModel extends MalcolmModel {
+public class EPICSv4ExampleModel extends MalcolmModel {
 	
 	// TODO extend this model with additional fields
