+++ org.eclipse.scanning.example/src/org/eclipse/scanning/example/malcolm/IEPICSv4Device.java
@@ -1,0 +1,16 @@
+package org.eclipse.scanning.example.malcolm;
+
+import java.util.Map;
+
+import org.epics.pvdata.pv.PVStructure;
+
+public interface IEPICSv4Device {
+
+	public void start() throws Exception;
+	
+	public String getRecordName();
+
+	public void stop();
+
+	public Map<String, PVStructure> getReceivedRPCCalls();
+}
